stm32: Add casts when printing small integers.
All these arguments are of type `mp_{u,}int_t`, but the actual
value is always a small integer. Cast it so that it can format
with the %d/%u formatter.

Before, the compiler plugin produced an error in the PYBD_SF6
build, which is a nanboxing build with 64-bit ints.

+++ ports/stm32/machine_uart.c
@@ -93,7 +93,7 @@ static void mp_machine_uart_print(const mp_print_t *print, mp_obj_t self_in, mp_
         #endif
         {
             mp_printf(print, "UART(%u, baudrate=%u, bits=%u, parity=",
-                self->uart_id, uart_get_baudrate(self), bits);
+                self->uart_id, uart_get_baudrate(self), (int)bits);
         }
         if (!(cr1 & USART_CR1_PCE)) {
             mp_print_str(print, "None");
